async working: cursors/reconnecting/rollback
Had to use gevent, but everything seems to be working. Cursors are good, reconnecting is good, Also fixed a bug where rollback was being called (possibly) inside the non-reentrant part of the protocol, where you can issue no commands until the current one is done.

This should pretty much transparently (gevent required) act the same as it did before, with the notable exception that it doesn't lock the entire program during a database query, not responding to input or signals. Also some debugging stuff... it'll say how long it's been trying to do a query, and what query it is.